Reorganise files in model serving metrics folder
