+++ frontend/src/pages/modelServing/screens/metrics/bias/BiasConfigurationBreadcrumbPage.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import { useOutletContext } from 'react-router-dom';
 import { getInferenceServiceDisplayName } from '~/pages/modelServing/screens/global/utils';
+import { GlobalModelMetricsOutletContextProps } from '~/pages/modelServing/screens/metrics/GlobalModelMetricsWrapper';
 import BiasConfigurationPage from './BiasConfigurationPage';
-import { GlobalModelMetricsOutletContextProps } from './GlobalModelMetricsWrapper';
 
 const BiasConfigurationBreadcrumbPage: React.FC = () => {
   const { model, projectName } = useOutletContext<GlobalModelMetricsOutletContextProps>();
