+++ frontend/src/pages/modelServing/screens/metrics/bias/BiasConfigurationEmptyState.tsx
@@ -7,7 +7,10 @@ import {
   Title,
 } from '@patternfly/react-core';
 import { WrenchIcon } from '@patternfly/react-icons';
-import { EMPTY_BIAS_CONFIGURATION_DESC, EMPTY_BIAS_CONFIGURATION_TITLE } from './const';
+import {
+  EMPTY_BIAS_CONFIGURATION_DESC,
+  EMPTY_BIAS_CONFIGURATION_TITLE,
+} from '~/pages/modelServing/screens/metrics/const';
 
 type BiasConfigurationEmptyStateProps = {
   actionButton: React.ReactNode;
