+++ frontend/src/pages/modelServing/screens/metrics/bias/BiasConfigurationTableRow.tsx
