+++ frontend/src/pages/modelServing/screens/metrics/performance/ServerMetricsPage.tsx
@@ -1,7 +1,7 @@
 import { PageSection, Stack, StackItem } from '@patternfly/react-core';
 import React from 'react';
 import MetricsPageToolbar from '~/pages/modelServing/screens/metrics/MetricsPageToolbar';
-import ServerGraphs from '~/pages/modelServing/screens/metrics/ServerGraphs';
+import ServerGraphs from '~/pages/modelServing/screens/metrics/performance/ServerGraphs';
 import { ServerMetricType } from '~/pages/modelServing/screens/metrics/ModelServingMetricsContext';
 import EnsureMetricsAvailable from '~/pages/modelServing/screens/metrics/EnsureMetricsAvailable';
